Modified behaviour of ParameterisedHolder::shouldSave, which controls which files are attributes are saved to file.
It's a small change, but one that we're a bit nervous about, because it potentially affects a bunch of things.  See Changes, comments, and new unit test for more details.

Summary is: attributes on the ParameterisedHolder which are not in the Cortex parameter mapping are only written to file if they have changed from default ( this matches the standard Maya behaviour ).

Changed to fix issue where Arnold parameters started being saved into scenes where Arnold was not used.